TST: Move some consistency expanding tests to test_expanding
